added dark mode & renamed some components
