+++ src/components/Artists/ArtistSearchSection.tsx
@@ -1,10 +1,10 @@
 import { ChangeEvent, FormEvent, useState } from "react";
 
-type ArtistSearchFormProps = {
-  handleSearch: (searchKey: string) => Promise<void>;
+type ArtistSearchSectionProps = {
+  handleArtistSearch: (searchKey: string) => Promise<void>;
 };
 
-const ArtistSearchForm = ({ handleSearch }: ArtistSearchFormProps) => {
+const ArtistSearchSection = ({ handleArtistSearch }: ArtistSearchSectionProps) => {
   const [searchKey, setSearchKey] = useState("");
 
   function handleChange(e: ChangeEvent<HTMLInputElement>) {
@@ -13,7 +13,7 @@ const ArtistSearchForm = ({ handleSearch }: ArtistSearchFormProps) => {
 
   function handleSubmit(e: FormEvent<HTMLFormElement>) {
     e.preventDefault();
-    handleSearch(searchKey);
+    handleArtistSearch(searchKey);
   }
 
   const form = (
@@ -37,4 +37,4 @@ const ArtistSearchForm = ({ handleSearch }: ArtistSearchFormProps) => {
   return form
 };
 
-export default ArtistSearchForm;
+export default ArtistSearchSection;
