+++ src/components/LoginPage.tsx
@@ -1,4 +1,4 @@
-const Login = () => {
+const LoginPage = () => {
   const CLIENT_ID = "c418f8eeb1764167baf4279404c77cba";
   const REDIRECT_URI = import.meta.env.PROD ? "https://spotify-song-keyword-search.vercel.app/" :  "http://localhost:3000"
   const AUTH_ENDPOINT = "https://accounts.spotify.com/authorize";
@@ -15,4 +15,4 @@ const Login = () => {
   );
 };
 
-export default Login;
+export default LoginPage;
