+++ src/components/Main.tsx
@@ -1,5 +1,5 @@
-import ArtistSearchForm from "./ArtistSearchForm";
-import ArtistSection from "./Artists/ArtistSection";
+import ArtistSearchForm from "./Artists/ArtistSearchSection";
+import ArtistResultTable from "./Artists/ArtistResultTable";
 import { useState } from "react";
 import { Artist } from "../Types";
 import SongSection from "./Songs/SongSection";
@@ -8,13 +8,13 @@ export type SearchProps = {
   token: string;
 };
 
-const Search = ({ token }: SearchProps) => {
+const Main = ({ token }: SearchProps) => {
   const [artists, setArtists] = useState<null | Array<Artist>>(null);
   const [selectedArtist, setSelectedArtist] = useState<null | Artist>(null);
 
 
   /* TODO: add pagination so users can select between more than the top 5 artists that show up */
-  async function handleSearch(searchKey: string) {
+  async function handleArtistSearch(searchKey: string) {
     const params = new URLSearchParams({
       q: searchKey,
       type: "artist",
@@ -39,11 +39,11 @@ const Search = ({ token }: SearchProps) => {
   return (
     <>
       <section>
-        <ArtistSearchForm handleSearch={handleSearch} />
+        <ArtistSearchForm handleArtistSearch={handleArtistSearch} />
       </section>
       <section>
         {artists ? (
-          <ArtistSection
+          <ArtistResultTable
             artists={artists}
             setSelectedArtist={setSelectedArtist}
             selectedArtist={selectedArtist}
@@ -59,4 +59,4 @@ const Search = ({ token }: SearchProps) => {
   );
 };
 
-export default Search;
+export default Main;
